+++ test/bscdodoflash.test.ts
@@ -3,11 +3,11 @@ import { expect } from "chai";
 import { ethers } from "hardhat";
 import * as IERC20 from "../artifacts/@openzeppelin/contracts/token/ERC20/IERC20.sol/IERC20.json";
 import { DODOFlashloan, DODOFlashloan__factory } from "../typechain";
-import { dodoV2Pool, erc20Address } from "../constants/addresses";
+import { bscTokens, dodoV2Pool } from "../constants/bsc";
 import { Contract } from "@ethersproject/contracts";
 import { getBigNumber } from "../utils";
 
-describe("dodo flashloan on polygon", () => {
+describe("dodo flashloan on bsc", () => {
   let Sample: DODOFlashloan;
   let owner: SignerWithAddress;
   let addr1: SignerWithAddress;
@@ -29,7 +29,7 @@ describe("dodo flashloan on polygon", () => {
 
       await Sample.deployed();
 
-      DAI = new ethers.Contract(erc20Address.DAI, IERC20.abi, provider);
+      DAI = new ethers.Contract(bscTokens.DAI, IERC20.abi, provider);
     };
   });
 
@@ -38,19 +38,19 @@ describe("dodo flashloan on polygon", () => {
   });
 
   describe("DODO flashloan", async () => {
-    it("should execute flashloan", async () => {
-      // borrowing 1000 USDC from DODOs WETH/USDC pool
+    xit("should execute flashloan", async () => {
+      // borrowing 1 BUSD from DODOs ETH/BUSD pool
       await expect(
         Sample.dodoFlashLoan(
-          dodoV2Pool.WETH_USDC,
-          getBigNumber(1000, 6),
-          erc20Address.USDC
+          dodoV2Pool.ETH_BUSD,
+          getBigNumber(1),
+          bscTokens.BUSD
         )
       )
         .emit(Sample, "checkBorrowedAmount")
-        .withArgs(erc20Address.USDC, getBigNumber(1000, 6))
+        .withArgs(bscTokens.BUSD, getBigNumber(1))
         .emit(Sample, "payBackLoan")
-        .withArgs(erc20Address.USDC, getBigNumber(1000, 6));
+        .withArgs(bscTokens.BUSD, getBigNumber(1));
     });
   });
 });
