+++ test/uniswap/uniswap.test.ts
@@ -1,15 +1,15 @@
 import { SignerWithAddress } from "@nomiclabs/hardhat-ethers/signers";
 import { expect } from "chai";
 import { ethers } from "hardhat";
-import { ERC20Mock, UniswapFork, UniswapFork__factory } from "../typechain";
+import { ERC20Mock, UniswapFork, UniswapFork__factory } from "../../typechain";
 import {
   USDC_WHALE,
   erc20Address,
   uniswapRouter,
-} from "../constants/addresses";
-import { getErc20Balance, impersonateFundErc20 } from "../utils/token";
-import { getPriceOnUniV2 } from "../utils/v2/getPrice";
-import { getBigNumber, getERC20ContractFromAddress } from "../utils";
+} from "../../constants/addresses";
+import { getErc20Balance, impersonateFundErc20 } from "../../utils/token";
+import { getPriceOnUniV2 } from "../../utils/v2/getPrice";
+import { getBigNumber, getERC20ContractFromAddress } from "../../utils";
 
 describe("Swap on uniswap fork on polygon", () => {
   let Fork: UniswapFork;
